+++ CHANGELOG.rst
@@ -1,6 +1,10 @@
-^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-Changelog for package catkin_tools
-^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+^^^^^^^^^
+Changelog
+^^^^^^^^^
+
+Forthcoming
+-----------
+* parallel build style initialisation
 
 0.4.3 (2017-01-05)
 ------------------